API/requirements.txt
fastapi
uvicorn
torch
tensorflow
numpy
pillow
